requirements.txt
aws-cdk-lib==2.43.1
python-dotenv==0.21.0
constructs==10.1.114
boto3==1.24.81
botocore==1.27.81
